fix: use TestKit for StrictSerialization tests and update API approval
- Move StrictSerializationSpecs to Akka.Hosting.TestKit.Tests project
  using the standard TestKit pattern (follows HostingSpecSpec.cs)
- Update CoreApiSpec.ApproveCore.verified.txt with new
  WithStrictSerialization API surface
